#2135: second draft done
