+++ src/frontend/src/pages/Calendar/CalendarPage.tsx
@@ -14,6 +14,7 @@ import {
   EnumToArray,
   calendarPaddingDays,
   daysInMonth,
+  exampleDesignReview1,
   testDesignReview1
 } from '../../utils/design-review.utils';
 import ActionsMenu from '../../components/ActionsMenu';
@@ -25,7 +26,7 @@ const CalendarPage = () => {
 
   const EventDict = new Map<Number, DesignReview[]>();
   // TODO remove during wire up ticket
-  EventDict.set(new Date().getDate(), [testDesignReview1]);
+  EventDict.set(new Date().getDate(), [exampleDesignReview1]);
   EventDict.set(new Date().getDate() + 3, [testDesignReview1, testDesignReview1]);
   EventDict.set(new Date().getDate() + 4, [testDesignReview1, testDesignReview1, testDesignReview1]);
   const designReviewData: DesignReview[] = [testDesignReview1, testDesignReview1];
