+++ src/frontend/src/pages/Calendar/SummaryModal/SummaryComponents/DesignReviewDelayModal.tsx
@@ -1,12 +1,11 @@
 import { TextField, Link, FormLabel, FormControl } from '@mui/material';
 import { useState, ChangeEvent } from 'react';
 import { DesignReview, wbsPipe } from 'shared';
-import { useToast } from '../hooks/toasts.hooks';
-import { routes } from '../utils/routes';
-import NERModal from './NERModal';
 import { Link as RouterLink } from 'react-router-dom';
-import { NERButton } from './NERButton';
-import NERSuccessButton from './NERSuccessButton';
+import NERModal from '../../../../components/NERModal';
+import NERSuccessButton from '../../../../components/NERSuccessButton';
+import { useToast } from '../../../../hooks/toasts.hooks';
+import { routes } from '../../../../utils/routes';
 
 export const DesignReviewDelayModal: React.FC<{ open: boolean; onHide: () => void; designReview: DesignReview }> = ({
   open,
